Project Structure Updated

+++ project3_read_trace/cpp_files/sim_proc.h
@@ -2,9 +2,9 @@
 #define SIM_PROC_H
 
 typedef struct proc_params{
-    unsigned long int rob_size;
-    unsigned long int iq_size;
-    unsigned long int width;
+    int rob_size;
+    int iq_size;
+    int width;
 }proc_params;
 
 // Put additional data structures here as per your requirement
